L-fix Assignment3

+++ src/lessons1/Assignment3.java
@@ -4,8 +4,8 @@ public class Assignment3 {
 
 	public static void main(String[] args) {
 
-        for(int i=1;i<=5;i++){
-            for(int x=1 ; x<=i;x++){
+        for(byte i=1;i<=5;i++){
+            for(byte x=1 ; x<=i;x++){
 
                 System.out.print(x);
